SUMMARY:  fix bug that occurs when some rows have no hyperedges, add                   a debugging function      CHECKLIST:
     1)   Did you add any source files? no

     2)   Did you update or change the build system?  no

     3) Did you rerun the test suite on at least the packages that you
        changed?  yes, part of it

If a process is part of a row that has no edges, then skip doing the allreduce
across the row to get the correct edge weight.  It involves giving
NULL buffers to MPI_Allreduce which crashes some MPI implementations.

I added a function for debugging that prints out a zoltan_pins structure.
I'm leaving it in commented out because it was useful.