SUMMARY: Minor changes for BJAM XXX_config.h files.
BUGS: 3954


CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? No

2) Did you update or change the build system (other than just adding
   source files)? No

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes

4) If justified, did you type 'make partial-clean' and then rebuild all
   libraries and executables with 'make everything'? No, I only rebuilt
   from within SIERRA

5) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? No, I only ran Aria + Trilinos
   Dev test suite

COMMENTS:

There are just some minor changes I made in the course of getting the new
XXX_config.h files to work with SIERRA.  I am not sure these changes are
needed at this point but these files are cleaned up some.

The most notable change was to AnasaziConfigDefs.hpp in that I removed the
ifdefs and just unconditionally included the standard C++ headers.  We had
decided as a Trilinos leaders group that we would do this for all Trilinos
packages but it looks like Anasazi was not updated for this.

Actally, I am in the process of doing a fresh rebuild and test for the serial
version of Trilinos tests and examples right now but I don't want to wait to
check in so that the Aria + Trilinos Dev nightly tests will be sure to pass.