SUMMARY:   Adding <cmath> to the list of headers to be included
     BUGS:      None.

     CHECKLIST:

     1)   Did you add any source files?  No.

     2)   Did you update or change the build system?  No.

     3) Did you rerun the test suite on at least the packages that you
        changed?  Yes.

     COMMENTS:

     Some comments submitted by an external user who is generating Trilinos RPMs
mentioned the requirement of the <cmath> header for the compilation of Teuchos.
This header was not being added in the Teuchos_ConfigDefs.hpp file even though it
was being checked for by the autoconfiguration script.  This commit simply adds
"#include <cmath>" to the list of headers to be included by Teuchos.