SUMMARY:  Adding another non-symmetric eigensolver method to LAPACK wrappers (_GGEV/_GEGV)
BUGS:     None.

Roger has requested the addition of the _GGEV/_GEGV LAPACK wrappers to the
Teuchos::LAPACK class so that NOX can migrate to using the Teuchos::LAPACK
wrappers.  Because of a deprecation issue, it is necessary to check whether
the LAPACK library has _GGEV or _GEGV, which requires an additional m4
macro (acx_ggev.m4) for configuration.  This check was added to the configure.ac
and the package was re-bootstrapped.  Trilinos has been built and tested after
this change to make sure nothing bad has happened in adding this new wrapper.