Subject:  Changing Amesos KLU and Paraklete solvers over to use the new version of SuiteSparse.
BUG #: 3996

The SuiteSparse v3.0 library has been added to Amesos as an update to the current
UFsparse library that contained KLU and Paraklete.  This library has been modified
to change the filenames and function names so that headers and the library symbol
table will not create conflicts with other 3PLs that use elements of the SparseSuite
library (AMD, BTF, CHOLMOD, etc.).

In this check-in, the Amesos classes that interface with KLU and Paraklete are being
updated to use the new versions of these codes in the SuiteSparse directory.
The Makefile.export.amesos.in is also being updated so that users of Amesos get
the correct directories for the new SuiteSparse codes.

I have run the test harness on serial and parallel builds of Trilinos using the new
Amesos SuiteSparse library and have had no issues passing the tests.  Here's to hoping
for a smooth transition!