SUMMARY:   Cleaning up Amesos
     BUGS:      4040

     CHECKLIST:

     1)   Did you add any source files?  No.

     2)   Did you update or change the build system?  Yes.

     3) Did you rerun the test suite on at least the packages that you
        changed?  Yes.

     COMMENTS:

     "using namespace std" has been removed from Amesos_ConfigDefs.h and any
classes that should be in the standard namespace have been prefixed by "std::".
This change touched a lot of files in the Amesos repository, but should be relatively
harmless.

Also, the PaStiX library interface has been removed.  The interface files have
been empty since their check-in, so no interface really exists.  However, it was possible
to enable the interface through Amesos, which would have resulted in many compile-time
errors.  I am removing any reference to PaStiX in Amesos for the time being so users
do not get confused.

(checklist version 2007-06-20)