SUMMARY:   Modified test to allow modification of OrdinalType
     BUGS:      4049

     CHECKLIST:

     1)   Did you add any source files?  No.

     2)   Did you update or change the build system?  No.

     3) Did you rerun the test suite on at least the packages that you
        changed?  Yes.

     COMMENTS:

     The Teuchos::BLAS class has a templated version of the reference BLAS.
While this has existed for some time, it has not been tested, nor was their
a way to test it for OrdinalTypes other than int.  The BLAS test has a typedef
for the OrdinalType, but it is ignored in the proceeding code.  This check-in
modifies the test to be dependent upon the OrdinalType defined by OType.

NOTE:  When OType is changed to something other than "int", the test will not
compile.  This is due to an inconsistency between the Teuchos::BLAS class and
the BLAS wrappers.

(checklist version 2007-06-20)