SUMMARY:   removed test causing -pedantic warnings
     CHECKLIST:

     1)   Did you add any source files?  noe

     2)   Did you update or change the build system?  no

     3) Did you rerun the test suite on at least the packages that you
        changed?  yes

     COMMENTS:

In one fo the Array methods, there was a check of non-negativity on the index.
This index was of type size_t, which is defined by my computer (and by the standard)
to be 'unsigned' int of some sort. as a result, the check of non-negativity was guaranteed
to pass, which caused my compiler (gcc 4.0.1) to issue a warning under -pedantic.
I commented this test out.

(checklist version 2007-06-20)