SUMMARY: Fixing broken UnitTest test
BUGS: 4150

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? No

2) Did you update or change the build system (other than just adding
   source files)? No

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes

4) If justified, did you type 'make partial-clean' and then rebuild all
   libraries and executables with 'make everything'? No, just teuchos.

5) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? Yes, but only teuchos.

COMMENTS:

In version 1.4.2.1 of Teuchos_UnitTestRepository.cpp, Chris fixed a few
problems.  However, this broke the behavior of one of the tests where "*"
should match any test name.

I am fixing this by adding the code block:

  if (fullMatchStr == "*") {
    return true;
  }

This makes the failing test pass.

Note that I only tested teuchos and not all of Trilinos.  However, I believe
this is fine given the nature if the test case and the change in the code.
Also, there are very good unit tests for the unit testing framework :-)

With this change, all of the test cases run on gabriel.sandia.gov should all
pass.  I hope the same is true for all of the test cases on exetazo.