SUMMARY:   Fix to void * pointer passed to callbacks.
     BUGS:      NA

     CHECKLIST:

     1)   Did you add any source files?  no

     2)   Did you update or change the build system?  no

     3) Did you rerun the test suite on at least the packages that you
        changed?  yes

     COMMENTS:

We were passing a local variable as the void *data pointer to the
callbacks.  But the local variable was not always valid when the
callbacks were actually called; it had, sometimes, gone out of scope.
This caused a segfault for zdrive.inp.partkway-order-metis-v3, as the
ZOLTAN_OBJ_SIZE_FN was registered during migration with a local variable
in the migration routine, and then invoked in ordering after that local
variable had gone out of scope.

I made the void * pointer a global variable (just as the Mesh is) so that
its value would always be valid.

Oddly, this problem ran fine on paunchy under purify, but segfaulted on
both octopi and my mac (32- and 64-bit).

(checklist version 2007-06-20)