SUMMARY: Fixing bug in Teuchos::ArrayView copy constructor
BUGS: 4190

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? No

2) Did you update or change the build system (other than just adding
   source files)? No

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes

4) If justified, did you type 'make partial-clean' and then rebuild all
   libraries and executables with 'make everything'? Yes, both serial
   debug and mpi optimized.

5) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? Yes, all 940 serial debug
   an all but one of the 821 mpi optimized tests passed (see failing
   Sundance test below).

COMMENTS:

Here I have fixed a bug in the Teuchos::ArrayView copy constructor that
resulted in a debug assertion.  The bad code was realted to a raw pointer
node_ date member that was not being used.  I wrote a new test that
demonstrated the error (see ArrayView_test.cpp) and I fixed the problem by
removing all traces of the bogus node_ data member.

I will not merg this change back into the development branch since I have
already made significant changes to Teuchos::ArrayView there and this problem
does not exist anymore there.

NOTE: There is one failing test:

   Sundance - SimpleModelEvalTest.exe...                  ! FAILED      1 second

This test shows the output:

    ********************************
    *** Start of rSQP Iterations ***
    n = 25, m = 21, nz = 525

     k    f         ||c||s    ||rGL||s  QN ||Ypy||2 ||Zpz||2 ||d||inf alpha    time(sec)
     ---- --------- --------- --------- -- -------- -------- -------- -------- ---------
        0      0.28         -         -  -        -        -        -        - 0.007434845

    Total time = 0.007443 sec

    Oops!  Not the solution.  Some error has occured!

    Number of function evaluations:
    -------------------------------
    f(x)  : 6
    c(x)  : 6
    Gf(x) : 1
    Gc(x) : 2

    MoochoSolver: Caught an std::exception of type Thyra::Exceptions::IncompatibleVectorSpaces described as : ../../../../../../../../Trilinos/packages/thyra/adapters/epetra/src/Thyra_EpetraThyraWrappers.cpp:529\
    :

    Throw number = 1

    Throw test that evaluated to true: !isCompatible

    Thyra::get_Epetra_MultiVector(map,mv)

    Error, the following vector spaces are not compatible:

     *epetra_vs : Thyra::DefaultSpmdVectorSpace<double>{globalDim=21,localSubDim=10,localOffset=0,comm=Teuchos::MpiComm<int>{size=2,rank=0,rawMpiComm=0x1f6a180}}

     *mv->range() : EpetraVectorSpace[dim=21, local=10]

This is obviously code that should never run correctly in parallel so I don't
know it would ever pass in a parallel build.

This is concerning but note that when this test is run on exetazo, as show at:

    http://software.sandia.gov/trilinos/developer/test_harness/results/test_output.html?mid=61&bid=781&tid=98

it reports no output at all!

I think there might be something wrong with this test in parallel.

This test also produces no output on my linux machine gabriel as shown at:

    http://software.sandia.gov/trilinos/developer/test_harness/results/test_output.html?mid=62&bid=13&tid=39

Looking at Sundance/test/definition, I note that the tests just rely on the
return value which is unreliable in general.

I don't know if this test is really working or not.

Given all of this, I am going to check in, even with this Sundance error because:

1) The nature of the change in Teuchos::ArrayView makes me doubt that this
would case this type of error.

2) This sundance test does not produce any output on any of the tests runs
reported on the Trilinos test results page.  Also, the nature of the exception
tells me that the test should never be run in MPI mode in the first place.