Summary: Cmakified EpetraExt and other small improvements
Bugs: 4217, 4218

Here is what I did:

(*) Cmakified EpetraExt: This involved more work that Triutils because there
are a lot more files and a lot of conditional logic.  I did not add all of the
tests that get built but I did add all of them that get run (see bug 4217).
Even with all of this, it only took about three hours to do this.

(*) packages/epetraext/test/definition: Changed to avoid running the
MatrixMatrix test twice in serial :-) (see bug 4218)

(*) linux-gabriel-gcc-3.3.4-dbg-cmake, linux-gabriel-gcc-3.3.4-mpi-cmake:
Added conditional enables to try to build more EpetraExt code.

(*) Trilinos_Add_Test.cmake: Fixed a bug where a serial test was being added
in MPI mode when it should not be.

(*) TrilinosGlobalHelpers.cmake:
- Changed package out-of-order error to a fatal error
- Changed output message

(*) EpetraExt_HDF5.cpp: Changed the formatting of the ifdefs to see what was
going on.  I technically did not build the file but the changes I think are
very minor and safe so I feel good about this.

(*) CMakeLists.txt: Added EpetraExt

I tested this with serial debug and mpi optimized with cmake