SUMMARY:   finished Belos/Tpetra adapter
     CHECKLIST:

     1)   Did you add any source files?  no

     2)   Did you update or change the build system?  no

     3) Did you rerun the test suite on at least the packages that you
        changed?  no

     COMMENTS:

Finished Belos/Tpetra adapter, filling in MvScale() and MvPrint() routines in MultiVector traits class.
While this is technically not sound software engineering, as these routines are not actually used anywhere
in the current Belos/Tpetra tests, these are one-line wrappers which are identical to those in the Anasazi/Tpetra
adapters, so I am 99.99% sure that everything is okay.

(checklist version 2007-06-20)