SUMMARY:   I added unit tests for the Hermite Interpolator while getting      ready to refactor the InterpolatorBase class to implement a Cubic Spline      Interpolator.
     BUGS:      3984

     CHECKLIST:

     1)   Did you add any source files?  YES
     1.a) Did you do a 'cvs add'?  YES
     1.b) Did you add them to a Makefile.am file in some way?  YES
     1.c) Did you do a new bootstrap?  YES

     2)   Did you update or change the build system?  NO

     3) Did you rerun the test suite on at least the packages that you
        changed?  YES

     COMMENTS:

While adding the unit tests for the Hermite Interpolator, I discovered that the
derivatives coming out of it were wrong.  I poured over the calculations again
and discovered a mistake in the Hermite derivative calculations.  Once this
mistake was corrected, the unit tests verify that the Hermite interpolator is
exact for polynomials of degree three or less and is inexact for polynomials of
degree 4.  This is true for both x and xdot!

(checklist version 2007-06-20)