Summary: Finishing automatic XML export and HTML display of package dependencies
This checkin basically completes the automatic export of package dependency
information and the ability to generate an html page that show the package
dependencies.

Now, by default, when the user runs cmake to configure Trilinos, the
dependencies file:

  Trilinos/cmake/python/data/TrilinosPackageDependencies.xml

automatically gets updated.  Then, when the Trilinos developer checks in, they
will check in the updated dependency information automatically if they change
their package's dependencies.  If the dependencies are not changing, then this
file should not change.  I hope that the generation of this file will be the
same on all platforms but we will have to test that to see if that is true or
not.  This should make the dependency handling truly automatic and make it
hard not to maintain.

The XML file that is output can be changed by setting the variable
Trilinos_DEPS_XML_OUTPUT_FILE or its generation cna be turned off by setting
it to empty as:

  -D Trilinos_DEPS_XML_OUTPUT_FILE:STRING=""

By default, the html webpage is written to the build directory but this can be
changed by setting:

  -D Trilinos_DEPS_HTML_OUTPUT_FILE:STRING=""

I have also added some simple unit tests to make sure that we keep generating
the same XML and HTML files.  I have looked over the output carefully and I
think that the dependency handling is now 100% correct (after about the 5th
try).  I have verified that if you add or remove a package to the unit testing
list of packages, these tests fail.

Once again, these tests should not be run on MS Windows as they use bash
scripts.  We could change theses to python scirpts but then python would need
to be installed.

Other miscellaneous changes include:

(*) CMakeLists.txt: Changed Trilinos_PACKAGES_OVERRIDE to
Trilinos_PACKAGES_UNITTEST_OVERRIDE.  I want to create another list of
packages to prototype all of the Trilinos packages.

(*) cmake/TODO: Added a few more helpers

(*) sampleScripts/generate-dependencies-html-cmake: Adding a cmake configure
script to only output the Trilinos dependencies HTML file.

I have tested all of this by building from scratch with cmake both with serial
debug and mpi optimized.