SUMMARY:   I added testing for changing the nodes underneath the CubicSplineInterpolator.
     BUGS:      3948

     CHECKLIST:

     1)   Did you add any source files?  NO

     2)   Did you update or change the build system?  NO

     3) Did you rerun the test suite on at least the packages that you
        changed?  YES

     COMMENTS:

I found a bug in the DataStore clone operation and fixed it and added unit tests covering it.
In debug mode, the CubicSplineInterpolator will copy the nodes passed in
through setNodes and verify that they have not changed when interpolate is
called.  I added several unit tests to cover these cases.

(checklist version 2007-06-20)