SUMMARY:   Changing Epetraext-Zoltan interface to use native Zoltan library
     BUGS:      4266

     CHECKLIST:

     1)   Did you add any source files?  No.

     2)   Did you update or change the build system?  Yes.
     2.a) Did you remember to do a bootstrap and do a 'cvs add' on the
          Makefile.am files and the Makefile.in files?  Yes.
     2.b) [Optional but recommended] Did you blow away the libraries and
          executables and then rebuild?  Yes.

     3) Did you rerun the test suite on at least the packages that you
        changed?  Yes.

     COMMENTS:

     The EpetraExt-Zoltan interface has historically expected Zoltan to be
provided as a 3PL.  However, now that Zoltan is integrated into Trilinos,
the configuration of EpetraExt should try to use the native Zoltan if the
EpetraExt-Zoltan interface is enabled.
     This check-in changes the configure.ac file so that it checks for
--enable-zoltan if EpetraExt is configured with --enable-epetraext-zoltan.
It also checks for --with-parmetis, since the EpetraExt interface to Zoltan
uses the ParMETIS partitioner by default.  If both configuration options
do not exist, then it treats Zoltan as a 3PL.