Summary: Getting build to work on Darwin machine joda with LAM MPI
This was a hard port to do.  On this machine, for some reason, the C++
compiler called from /Users/kddevin/code/lib-m64/lam-7.1.4/bin/mpic++
does not allow duplicate libraries on the link line.  This caused two
errors when linking.  First, the MPI libraries can not be linked in
twice which is what happens with the current FindTPLMPI.cmake module
when using the MPI compiler (which already includes the libraries
once).  To get around this, I just had to set TPL_MPI_LIBRARIES to ""
and that solved that problem.  However, the other problem was that the
logic in PACKAGE_ADD_EXECUTABLE(...) was adding the DEPLIBS after the
other package libraries.  This resulted in CMake adding libisorropia.a
twice.  I fixed this by moving the DEPLIBS targets first and that
fixed that.  I also had to turn off Fortran.  While I could find a
Fortran compiler okay the automatic name mangling check did not work.
It will take some work to get the name mangling problem worked out but
I did not work to fight with this right now.

Also, I could not run the MPI tests because I could not figure out how
to get LAM MPI to run (I have never used LAM MPI before so I don't
know the magic).

Specific changes:

(*) cmake/package_arch/PackageAddExecutable.cmake: Moved DEPLIBS
libraries to the beginning of link line

(*) sampleScripts/darwin-mpi-release-isorropria-joda-cmake: Basic
configure options to get Isorropia and Zoltan to configure and build
on darwin machine zoda with LAM MPI.  Work needs to be done.

I have only tested this by building Isorropria and Zoltan and not
running anything.  I will immediately build on a Linux machine and see
what happens after this checkin.