Summary: Fixing AztecOO dependence on NOX/Thyra adapter tests
Bugs: 4239

The MPI_OPT_SHARED nightly build failed because AztecOO is disabled
and the NOX/Thyra tests did not respond correctly to having AztecOO
disabled.

(*) nox/test/epetra/Thyra_2Dsim/CMakeLists.txt: Added if test for
Stratimikos_ENABLE_AztecOO because that must be true for this test to
work.

(*) nox/test/epetra/Thyra_2Dsim/Thyra_2Dsim.C: Removed "AztecOO.h"
include because this code does not need it.  If AztecOO is not
enabled, this file can still be compiled, but the test will not be
run.

Tested on godel.sandia.gov with:

- Just did a rebuild of NOX and ran NOX tests after this change.