Summary: Adding --not-unit-test option to unit testing infrastructure
Bugs: None filed

Here, I just added the command-line option --not-unit-test to the unit
testing infrastructure.  This is going to allow me to disable single
tests (or globs of tests).  I added a test case to make sure it is
correct.  I will use this to disable the single troublesome
Tpetra_CrsMatrix test on gabriel.sandia.gov.  I only rebuild Teuchos
and reran the Teuchos test suite.  Given the nature of the change, I
think this is okay because I have very good unit tests for the
UnitTestRepository class.


Tested on godel.sandia.gov with:

- CMake serial debug rebuild of only Teuchos