SUMMARY:   I hacked in a fix for SinCosModel to pass nightly testing.
     BUGS:

     CHECKLIST:

     1)   Did you add any source files?  NO

     2)   Did you update or change the build system?  NO

     3) Did you rerun the test suite on at least the packages that you
        changed?  YES

     COMMENTS:

I'm using the Thyra::linearOpWithSolve non-member constructor to create a
Thyra::LinearOpWithSolveBase object for the SinCosModel's create_W function.
Under the hood, I'm creating a MultiVector and passing it off as a LinearOpBase
object.  Unfortunately, the linearOpWithSolve function ultimately ends up
trying to factorize the matrix _at initialization_!  This actually doesn't fail
catastrophically in debug mode because all the entries are filled with nan.
But in optimized mode, the memory is uninitialized and often ends up being
close to zero, so the rank of the matrix is determined to be zero and the
factorize throws an exception.  This is why I've been getting failed tests
during nightly testing.

I also added a feature to the SinCosModel to turn off nominalValues for testing.

And I simply updated the TODO with the fact that one of the items is completed.

(checklist version 2007-06-20)