SUMMARY:   Eliminate compile error when TEUCHOS_DEBUG enabled.
     BUGS:      4318

     CHECKLIST:

     1)   Did you add any source files?  No.

     2)   Did you update or change the build system?  No.

     3) Did you rerun the test suite on at least the packages that you
        changed?  Yes, with debug enabled.

     COMMENTS:

     The MyMultiVec class now has an overloaded function for operator<<,
this takes care of the compile error that was being seen with the cxx_mattest_complex.cpp
file.  Also, I moved the creation of ONE and ZERO into the appropriate functions
instead of having a global ONE and ZERO.  This is because ONE and ZERO are often
used in Anasazi's templated classes, so a global one generates warnings.

(checklist version 2007-06-20)