Several changes:
1)  Added Fad tests for inverse hyperbolic functions.  These aren't actually
part of the standard library, so I had to add implementations for float/double.

2)  Added a simple forward AD class that doesn't use expression templates.
This is primarily for testing to see what difference expression templates
makes.

3)  Added a fairly significant performance test that does a finite
element residual and Jacobian calculation.  It tests all of the various
Fad versions in a fairly realistic setting.  It also tests ADOL-C and the
simple forward AD class mentioned above.

4)  Made a small improvement to the vector fad storage to improve efficiency.

5)  Fixed a shadow warning coming from the parameter library.