SUMMARY:   Attempting to remove strcasecmp and strncasecmp warnings.
     BUGS:      NA

     CHECKLIST:

     1)   Did you add any source files?  no

     2)   Did you update or change the build system?  no

     3) Did you rerun the test suite on at least the packages that you
        changed? builds only

     COMMENTS:

We had tests for __STDC__ surrounding #includes of string.h and strings.h.
According to websites, it appears to be OK to include both.  So I did.
But I tested the compilation on only Mac and Linux so far; I'll need to
take it to other systems to be sure that it is OK.

(checklist version 2007-06-20)