Adding AztecOO and Galeri as optional test libraries.
For some reason building the examples is dependent upon AztecOO and Galeri,
but not the tests even though many of the tests also use Galeri and AztecOO.
When I tried to build Trilinos with Ifpack enabled, but not AztecOO
or Galeri the build failed for the tests.  So, I have added the cmake logic
into the tests that require AztecOO and/or Galeri so these build failures don't
occur.