SUMMARY:   StepperValidator initialization state testing
     BUGS:      4377

     CHECKLIST:

     1)   Did you add any source files?  NO

     2)   Did you update or change the build system?  NO

     3) Did you rerun the test suite on at least the packages that you
        changed?  YES

     COMMENTS:

I added stepper initialization state testing to the StepperValidator.  To get
the tests to pass, the steppers had to be updated a little.  During this
update, I fixed a few things like default initialization of data for the
steppers and a few other classes.  This commit also converts Implicit RK over
to require an intial condition, which meant I needed to change some of its
tests, including the convergence testing.

(checklist version 2007-06-20)