Allow overriding and nullifying TPL library names list from user cache
Here, I have changed the logic in the TPL_DECLARE_LIBRARIES(...) function to
allow the user to wipe out the list of libaries to search for a particular
TPL.  This is needed for the use case:

1) Support for the TPL needs to be enabled but the actually TPL library
functions are bundled with another library and you can't list the library more
than once.

This will be used by Brent to decare pthreads to be an offical TPL but allow
the pthreads library to be wiped out when building with MPI that already
includes pthreads.  There is one Mac machine that does not allow the same
library to be listed more than once.

(*) cmake/package_arch/TPLDeclareLibraries.cmake
- Added support for wiping out all of the libaries for a TPL with libraries
- Added extra print statements to show what libraries are being looked for and
what the user-definded search directories are

(*) sampleScripts/linux-gabriel-gcc-4.2.0-serial-debug: Changed the TPL find
varibles for MA28 to allow testing of this new TPL handling behavior.  It is
more verbose but is not too much so.

I tested this with with enabling MA28 a few different ways and manually
verified that things are getting set correctly.