SUMMARY:   Provide access to ColorMap for columns
     BUGS:

     CHECKLIST:

     1)   Did you add any source files? no

     2)   Did you update or change the build system?  no

     3) Did you rerun the test suite on at least the packages that you
        changed?  Yes

     COMMENTS:

     Now Isorropia offers a way to export "column" colors in a MapColoring object.

The way to do this is to use :
Teuchos::RCP<Epetra_MapColoring> Colorer::generateColMapColoring();
or
Teuchos::RCP<Epetra_MapColoring> Colorer::generateRowMapColoring(); for the row.

The previous generateMapColoring() method return row colors and is now declared as
deprecated.

By the way, these functions perform the coloring if the user has not done it yet.

Thanks to A. Salinger for the code.

(checklist version 2007-06-20)