Fixed data extraction in Teuchos::fromStringToArray when input strings have spaces
BUG # 4458

Summary:

Per the discussion in bug 4458, the Teuchos::fromStringToArray has issues when
the resulting Array is a container of std::string that have spaces.  For instance:

	Teuchos::Array<std::string> originalArray(3);
        originalArray[0] = "Option\t1";
        originalArray[1] = "Option\n2";
        originalArray[2] = "Option 3";

        std::string potentialArray = originalArray.toString();
        Teuchos::Array<std::string> malformedArray = Teuchos::fromStringToArray<std::string>(potentialArray);

This code generates a Teuchos::Array with strings that has different kinds of white spaces or newlines.
These strings are stored and then written out correctly in the call toString().  However, fromStringToArray()
reads in this string and generates the Teuchos::Array:

{ "Option", "Option", "Option" }

This is because of the extraction from the istringstream using the '>>' operator.
This check-in commits a modification to Teuchos::Array, introducing a templated function that performs
the extraction from istringstream.  For anything other than std::string, the extraction is
the same as before.  But for std::string, it grabs the unformatted string from the istringstream and
trims the white space before putting it in the Teuchos::Array.

I have added a unit test to test this issue and have ensured that all the previous
unit tests for the fromStringToArray() function still pass.