SUMMARY:   Made LB_Eval_HG always evaluate w.r.t. the hypergraph                 model, regardless of partitioning method used.
     BUGS:      4490

     CHECKLIST:

     1)   Did you add any source files?  no

     2)   Did you update or change the build system?  no

     3) Did you rerun the test suite on at least the packages that you
        changed?  small tests done; suite beginning.

     COMMENTS:

Recent code changes make the hypergraph construction depend on the
LB.Method when graph callbacks were used.  This was bad when LB.Method
was, say, RCB; we might register graph callbacks to allow evaluation
but Zoltan_LB_Eval_HG didn't know what to do with LB.Method==RCB.

For Zoltan_LB_Eval_HG, we always want the HYPERGRAPH model to be used;
that is, we always want to use_all_neighbors of a vertex in creating a
hyperedge.  I now pass in the model as an argument to the build function,
rather than testing LB.Method.  Zoltan_LB_Eval_HG always passes HYPERGRAPH
as the model.


(checklist version 2007-06-20)