SUMMARY:  Changing order in how numEvecs is computed.
     BUGZILLA BUG #(s):  4545

     CHECKLIST
     ---------

     1) Did you 'cvs add' all new files and directories [Y:N:N/A]?  N/A

     2) Did you add new source files to a CMakeLists.txt file [Y:N:N/A]?  N/A

     3)  Did you modify or add any source or build system files?  N/A


     DESCRIPTION OF CHANGES:

     The determination of how many eigenvectors the solver manager should
ask from the iteration was intertwined with the determination of if their are
conjugate eigenpairs and whether we need to keep one more eigenpair to avoid
splitting a conjugate pair.  So, first we determine whether we have split a
conjugate pair with the number of eigenvalues/vectors we are returning.
Then we determine if the number of converged eigenvectors are in the principle
set, or whether we need to ask for more vectors to get all the converged ones.

     Confusing, eh?

     Anyways, I do not know for sure if this will address the issue that Salinas
is seeing with BKS.  However, I have put many debugging statements around this
block of code to get more information if this doesn't happen to resolve their issue.


(checklist version 2009-07-14)