Fixed the trilinos-test cron driver, apparently I can't checkout individual files into a different directory like I hoped. Instead the scripts will just need to be run from a different subdir to separate release and dev testing scripts
Made similar changes to the other release testing platform as well as cleaned up
the cron script a bit to remove builds that were commented out and probably
won't be run on that machine any time soon.

These changes will need to be merged into the development branch as well.