SUMMARY:   Addition of new StatusTestOutputFactory class.
     BUGZILLA BUG #(s):  None.

     CHECKLIST
     ---------

     1) Did you 'cvs add' all new files and directories [Y:N:N/A]?  Y

     2) Did you add new source files to a CMakeLists.txt file [Y:N:N/A]?  Y

     3)  Did you modify or add any source or build system files?  Y
     3a) Did you rebuild the libraries and executables?  Y
     3b) Did you rerun the test suite on the packages that you changed and
         all packages that depend on the modified packages?  Y
     3c) Did you change the required or optional dependencies for any
         package?  N


     DESCRIPTION OF CHANGES:

     The logic used to determine which StatusTestOutput class is created and what options
are passed in to the generation of this class has now been moved into a factory.
This eliminates the need to replicate the logic in each of the solver managers.
     All the solver managers have been modified to use this new BelosStatusTestOutputFactory.
There is a new enumeration, Belos::OutputType, that determines which BelosStatusTestOutput
class to use.  That enumeration is passed in during the construction of the
factory and is hard coded to Belos::General right now.  This results in the old
chatty output behavior, which uses the StatusTestGeneralOutput class.  In the
next couple days, the output type will become a parameter list option for these
solvers, giving each solver access to the new BelosStatusTestResNormOutput class.

(checklist version 2009-07-14)