Added --no-eg-git-version-check
BUGS: 4645

I added the options --no-eg-git-version-check to allow users to turn off the
version check for eg and git.  This will be needed to use the checkin-test.py
script on the SCICO LAN.  Othwerwise, we expect Trilinos developers to be
using the official versions of eg/git (or they are asking for trouble).

I did not append the test results (to avoid clutter) but they all passed.