Add a xml-driven test-driver for Ifpack/AztecOO.
This test-driver is simply a conversion of the one already written for
Tifpack/Belos. Having both will be very valuable for comparing the
correctness and performance of Tifpack and Ifpack on matrices that are
read from matrix-market files. (A stub is in place for adding Harwell-
Boeing file support also, but not yet implemented.)