Changed the sort order of unit tests to maintain the natural order
Here I have changed the the sort order of the unit tests to sort by group name
first and then by the oder they where added.  If we can assume that a group is
isolated to a single source file, then the unit tests will remain sorted in
the same order they are added in a translation unit.  This makes it possible
to write unit tests that start very basic and then get more complex.

I also removed obsolute unit test ordering code