Put in logic to allow shadowing warnings when TBB is not enabled.
Here, I say that someone turned off shadowing warnings in Tpetra because TBB
has shadowing warnings.  What I did was to put in the logic to enable
shadowing warnings when Kokkos_ENABLE_TBB EQUAL TRUE.  However, when building
I found lots of shadowing warnings in Kokkos and Tpetra.  If fixed a few in
Kokkos but there are a *bunch* in Tpetra which I did not fix.  Therefore, I
had to turn back off shadowing warnings.

I would suggest that some Tpetra developer re-enable shadowing warnings in
tpetra/CMakeLists.txt (when TBB is not enabled) and get the shadowing warnings
cleaned up.  Note that SIERRA is moving to turn on -Wshadow with -Werror so we
need to clean up all shadowing warnings or we are not going to be able to
support SIERRA (or other customers that want to turn on -Wshadow -Werror).