Completed work on the short polynomial branch.
I finished getting propagation of short polynomials (i.e., polynomials
of degree less than the full expansion) working.  The primary use case
for this is for linear problems, and significantly speeds up the
computation.  In doing this work, I changed the way the sparse tensors
are allocated to make this more explicit so that the right size is
always generated.  This required changing every test and example for
Stokhos, and a few in Sacado and NOX.  I also added two NOX examples for
linear problems that use this capability.  I also added a constant
source function so that I can do common linear diffusion problems.
Finally I modified the * operation of CacheFad to eliminate extra zero
computations in some situations, which can significantly improve the
performance of PCE.  It turns out ELRFad is better at this, so in the
future I need to create an ELRCacheFad.