Adding testing for development branch on windows
Previously we were only testing the release branch on windows. We need to have
development branch testing to ensure we don't break windows builds.

initially these test builds will be experimental to give people a chance to fix
any regressions which may have crept in while we weren't testing. Eventually
these will be promoted to nightly tests.

The shared build right now only has a few packages converted over to use the
required dll export macro that windows builds rely on. Because of that the
package list is overridden to only included those packages instead of the normal
set of primary and secondary stable packages

The location of dlls needs to be on the path so they can be found when tests
are run.

Note that there are still some issues to be worked out for the testing, but
commit is a good start to getting windows development testing up and running