Changes to configuration and build of ITAPS support:
1) Mesquite previously compiled support for all ITAPS APIs w/out needing
an implementation of each API to be installed by keeping its own copy
of all of the ITAPS headers.  This changes the build system such that an
implementation of each API must be present for Mesquite to enable support
for that API, and Mesquite now uses the implementations' versions of the
ITAPS API headers rather than its own copy.  This change is
necessary because of the ITAPS Fortran name-mangling mess.  I'm tired of
tracking down issues such as the case where a user builds MOAB, then at a
later date installs a Fortran compiler and at a yet later date builds
Mesquite. This scenario results in a MOAB and Mesquite being unable to
link together because Mesquite did Fortran name-mangling for the iMesh
API and MOAB did not.

2) As a part of implementing the above change, change the configure options
from --enable-imesh, etc. to --with-imesh=<path>, where <path> must be the
iMesh-Defs.inc file that came with the iMesh implementation.  Also, honor
the Fathom group convention of using environmental variables such as
IMESH_DEFS to specify the name of that file. If IMESH_DEFS is defined, then
configure will default to enabling iMesh support using the specified file.
This behavior can be overridden by specifying --without-imesh.

3) Build the Mesquite support for ITAPS APIs as separate libraries (libmsqimesh,
libmsqigeom, etc.)  This way when Mesquite is built as a shared library, the
user need only link ITAPS API implementations (e.g. MOAB, CGM, etc.) if they
want to use the ITAPS APIs (the core Mesquite library doesn't depend on
the ITAPS API implementations.)