Split out EpetraExt_PyUtil into multiple files
Functionality that was in EpetraExt_PyUtil was split out into a
PyTrilinos_Util header/source combo and an Epetra_PyUtil header/source
combo.  The primary motivating factor was to make the function
convertEpetraOperatorToPython(), which attempts to downcast an
Epetra_Operator to a derived class, available to the Epetra module,
and ultimately to the NOX.Epetra module as it deals with
Epetra_Operators and cross-language polymorphism.  Functions that
dealt only with the Python/C API were placed in PyTrilinos_Util and
functions that dealt with Epetra objects were placed in Epetra_PyUtil.