Added CI testing for preCopyrightTrilinos packages (BUG: 4692)
Here I added the git calls and CMake code to also get the changed files in
preCopyrightTrilinos when determining what packages have changed and need to
be enabled.  Now when someone commits and pushes a change to a
preCopyrightTrilinos package, the CI testing will detect changes these these
SS packages and enable them for testing.

This was slightly more complex than I had first throught and took a full 30
minutes to write the code, test it (manually) and debug it.

With this full testing support for preCopyrightTrilinos should be finished.
Now any preCopyrightTrilinos package can be made SS.