Improved SG vector/multivector InArgs support.
I added subclasses of VectorOrthogPoly for
Epetra_Vector's and Epetra_MultiVector's that
automate the process of converting to/from
EpetraExt::Block{Multi}Vector's, and changed
In/OutArgs to use these.  I also added InArgs
for the SG basis, quadrature, and expansion 
objects to formalize how these are passed
through model evaluators.  I then took the
quadrature object out of the VectorOrthogPoly,
which was a bad place to put it.  Finally I 
added an "inverse" SG model evaluator that maps
SG In/OutArgs to EpetraExt::BlockVector's to that
the response-only model evaluators can be used
with SG problems.