Add generic cron_driver.sh, usable from any Linux or Mac box.
The bash shell script clones or updates the Trilinos repository
to get the most up to date scripts. Then it downloads and installs
the release of CMake. Then it uses that ctest to drive the script
TrilinosDriverDashboard.cmake.

That script configures/builds/tests the TrilinosDriver project at
cmake/ctest/drivers/CMakeLists.txt. Each *test* on the driving
dashboard is a full separate run of a Trilinos dashboard.

At most, one of the subdirectories of cmake/ctest/drivers is used
on a given machine.