Fixing ScalarTraits::isnaninf(...) on Intel 10.1 compiler
The isnaninf(...) test taken from NOX did not correctly flag Infs.  I
modified the rountines for float and double to use a new generic
template function that adds another test to compare to
std::numeric_limits::infinity().  This test will be a little more
expensive now but you should only use this test in production code on
a reduced quantity like a norm.

My guess is that the Intel 10.1 C++ compiler saw the 0.0
multiplication and just set it to 0.0 without checking that x was
actually NaN or Inf.