Catch a few more error conditions.
Queue an error if git is not found, or if GIT_EXE does not
exist.

If the source directory does not exist just prior to calling
ctest_update, then queue an error, report and return immediately.
(If not, ctest_update simply exits the script anyhow and we do
not get queued errors reported... This is a ctest bug that should
be fixed, but this is a workaround that avoids the problem for now.)

Always call REPORT_QUEUED_ERRORS() prior to any RETURN() calls in
the TRILINOS_CTEST_DRIVER function.

Explicitly avoid using RETURN_VALUE in ctest_submit calls.
The replacement wrapper macro for ctest_submit handles it.

Make calls to createDir in download-cmake.py safe for parallel
execution by adding exception handling and existence checks.