Added a callback interface to the preconditioner
This is intended as a replacement for the "request"
parameter interface.  It provides several benefits,
   1. Its more structured
   2. The preconditioner is more proactive about getting
      "extra" data
   3. Doesn't require complex nesting to get through
      a deep preconditioner hierarchy
The interface itself is based on the listener pattern
with a static dispatch to different data types.

Somethings remain to be implemented
   1. A Warning interface: tell the user this is
      what I need
   2. More sophisticated error handling
   3. Support for multigrid like hierachies: using
      ID field?
   4. Pass Jacobian with message?

The primary difficulty with this is that it forces
the user to implement some possibly intricate interfaces.
Hopefully the term "Callback" is familiar enough to
make that easy.