Update post-receive-email hook
Updated the post-receive-email hook to inspect the branch that was being pushed
to. If the branch begins with a package name eg "zoltan" then only that
package's checkin list will be notified of the changes. This should cut down on
the "spam" from when people merge master into their branches so that they are up
to date.

There are two notable exceptions to this rule pushes to the master branch and
the trilinos-release* branches will continue to notify based solely on the files
that were changed.

The comparison for determining what is a "package's branch" is that the prefix
of the registered email addresses(ie, epetra-checkins@software.sandia.gov is
registered, epetra would be the prefix) is compared to the lower case name of
the branch. If the branch starts with the same name as the prefix then the email
regarding that push is sent only to that email address. Example: Using the email
list mesquite-checkin@sandia.gov the package name will be determined to be
"mesquite". There is currently a branch "MESQUITE_2_0". The branch name is
converted to lower case "mesquite_2_0" and then searched to see if it begins
with "mesquite". Note that this method will not find branches that just have the
name of the package in them somewhere, the branch must begin with the package
name for this filtering of emails to happen.

If the branch in question isn't one of the special branches(master or
trilinos-release*) then the push email will be sent to the default mailing list
which is trilinos-checkins@software.sandia.gov. Currently no branch on the
repository would fall through to that email address.