Changed RequestHandler to a "contains" relationship
Changed RequestHandler to a "contains" relationship
not an "is-a" relationship. So now the PreconditionerFactory
contains a RequestHandler that it queries for requests. Also
modified the InverseLibrary so that it passes along its request
handler to all the factories it creates.  The InverseFactory
class also contains a request handler interface.

Some ramifications: This moves the InverseLibrary to be
more and more essential.  It already does much of the work
for the user.  However, now it gets piled on with handling
the call back interface. I view this as not such a big deal
as the functionality it provides and manages is pretty
substantial/useful.