Changing 'eg status' to 'eg diff --shortstat' for updated in eg/git
It would seem that with the upgrade to eg/git 1.7.0.4, eg status no
longer returns 1 if there are modified files.  Therefore, I had to
change the script to use 'eg diff --shortstat' to detect if there are
uncommitted changes before continuing.

I had to update the unit tests as well.  Actually, the use case where
there are modified files is not well unit tested.