Improving the outputting of non-deleted RCPNodes
Here, I have done two things.  First, I improved the message printed out at
the end of the program to suggest multiple options for the possible causes of
non-deleted RCPNodes.  It turns out that circular references may not actually
be the most common cause of this.  A more common cause is calling abort() or
exit().

I also changed the code to print out the RCPNodes in the order of
insertionNumber.  This should make it easier to see trends and is just a
better thing to do.  Before, the order was determined by multimap and some
hash key on void* pointer addresses.

These are all little steps to make it easier to debug problems (either
directly or indirectly) related with the usage of the Teuchos memory
management classes.

I added a static function to Teuchos::RCPNodeTracer in Teuchos_RCPNode.hpp to
remove some duplication in unit test software.  However, this will require
rebuilding everything that depends on this header which is much of Trilinos.
However, given the simple nature of the change and the high quality of the
unit test I don't think I need to test all of Trilinos in this case.

=============================
Build/Test Cases Summary
=============================

Thu May 13 14:11:32 MDT 2010

Enabled Packages: Teuchos

Build test results:
-------------------
0) MPI_DEBUG => passed: Trilinos/MPI_DEBUG: passed=57,notpassed=0 (4.63146315018 min)
1) SERIAL_RELEASE => passed: Trilinos/SERIAL_RELEASE: passed=54,notpassed=0 (2.92619298299 min)

Local commits for this build/test group:
----------------------------------------
10181c0 Improving the outputting of non-deleted RCPNodes