add ability to get contents of /proc/meminfo when application fails
zdrive defines a signal handler that will cause nodes to print out
/proc/meminfo on failure.  The new input file option:

zoltan memory debug level = n

sets the DEBUG_LEVEL in mem.c.  If it is 1 or greater, then on failure
the processes will try to print out /proc/meminfo on Linux nodes.
"Failure" is one of several signals, or it is a return value of
ZOLTAN_MEMERR from the load balancing function.