(From Sierra) New simple use case from Ulrich Kuettler in which a shared mesh entity is downgraded to a ghosted mesh entity.
Added to BulkData::modification_end() enforcement of the
invariant that shared mesh entities must be in the owned closure.
This invariant is enforced by destroying a modified formerly-shared
mesh entity that is no longer in the owned closure, and destroying
any ghost mesh entities that are dependent on that formerly-shared
mesh entity.