Added new diagonal scaling functionality
The idea is that perhaps a particular matrix
is easier to precondition if it is diagonally scaled.
Currently this method uses only right scaling (column
scaling). Included a simple unit test to excercise
this new functionality.

Also added a "describe" function to the PreconditionerLinearOp
class.  This is useful for debugging.