Added rcpCloneNode(p) function for RCP.
The rcpCloneNode(p) function is the right way to create a new private
reference count.  I added this function so that Charon could use it and it
should be used for other purposes as well.

This is only added a new helper function and unit tests so I did not feel that
is was necessary to test down-stream packages.

Build/Test Cases Summary
Enabled Packages: Teuchos
0) MPI_DEBUG => passed: passed=55,notpassed=0 (4.51 min)
1) SERIAL_RELEASE => passed: passed=52,notpassed=0 (3.74 min)