A computation protected by a degeneracy check was returning NaNs because the degeneracy check was failing.  This corrects the degeneracy check.
If two overlapping faces overlapped in a line segment, then the area of
the overlap was zero.  To guard against this computation, if the number of
nodes in an overlap polygon was less than three, the overlap was not
computed.  Unfortunately, if the polygon was degenerate, the number of
nodes in the polygon could exceed 2.  This is one way to ensure the number
of nodes in the polygon will not exceed 2 for degenerate overlap polygons.