(From Sierra) Refactoring of mesh modification end synchronization to propagate modification state of shared entities before processing resolving ghost entity clean up.
Enlarge 2D and 3D grids for unit testing deletion of an element.
Fix bugs in unit test to allow unit test executable to run up to 13 processes.

If a ghost entity changes to shared status due to mesh modification
proper handling is to destroy ghost communication but preserve the
entity.